backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.6,<5.0
django-filter>=24.0
psycopg2-binary>=2.9,<3.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0
celery>=5.4,<6.0
Pillow>=11.0,<12.0
python-decouple>=3.8,<4.0
drf-spectacular>=0.28,<1.0
gunicorn>=23.0,<24.0
